docs/examples/python-with-db/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.10
